# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NP Wallet - Web3 Wallet API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Wallet {
  id: number;
  name: string;
  address: string;
  networkId: number;
  color: string;
  /** @nullable */
  totalValueUsd?: number | null;
  /** @nullable */
  change24hPercent?: number | null;
  createdAt: string;
}

export interface WalletInput {
  name: string;
  networkId: number;
  color?: string;
  /** @nullable */
  importedAddress?: string | null;
}

export interface WalletUpdate {
  name?: string;
  color?: string;
}

export interface TokenHolding {
  id: number;
  walletId: number;
  tokenSymbol: string;
  tokenName: string;
  balance: number;
  valueUsd: number;
  priceUsd: number;
  logoUrl: string;
  change24hPercent: number;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  send: 'send',
  receive: 'receive',
  swap: 'swap',
  contract: 'contract',
} as const;

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  failed: 'failed',
} as const;

export interface Transaction {
  id: number;
  walletId: number;
  type: TransactionType;
  status: TransactionStatus;
  fromAddress: string;
  toAddress: string;
  amount: number;
  tokenSymbol: string;
  /** @nullable */
  toTokenSymbol?: string | null;
  /** @nullable */
  toAmount?: number | null;
  valueUsd: number;
  networkId: number;
  hash: string;
  /** @nullable */
  gasFeeEth?: number | null;
  /** @nullable */
  gasFeeUsd?: number | null;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface SendInput {
  toAddress: string;
  amount: number;
  tokenSymbol: string;
  /** @nullable */
  note?: string | null;
}

export interface SwapInput {
  fromTokenSymbol: string;
  toTokenSymbol: string;
  amount: number;
}

export interface Token {
  id: number;
  symbol: string;
  name: string;
  priceUsd: number;
  change24hPercent: number;
  logoUrl: string;
  marketCapUsd: number;
  volume24hUsd: number;
  isPopular?: boolean;
}

export interface Network {
  id: number;
  name: string;
  symbol: string;
  chainId: number;
  rpcUrl: string;
  explorerUrl: string;
  logoUrl: string;
  isTestnet: boolean;
  color?: string;
}

export interface Nft {
  id: number;
  walletId: number;
  name: string;
  collection: string;
  tokenId: string;
  imageUrl: string;
  networkId: number;
  /** @nullable */
  floorPriceEth?: number | null;
  /** @nullable */
  floorPriceUsd?: number | null;
  /** @nullable */
  traits?: string | null;
}

export interface PortfolioSummary {
  totalValueUsd: number;
  change24hUsd: number;
  change24hPercent: number;
  topAsset: string;
  walletCount: number;
  networkCount: number;
  totalTransactions?: number;
}

export interface PortfolioDataPoint {
  timestamp: string;
  valueUsd: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  send: 'send',
  receive: 'receive',
  swap: 'swap',
  contract: 'contract',
  nft: 'nft',
} as const;

export type ActivityItemStatus = typeof ActivityItemStatus[keyof typeof ActivityItemStatus];


export const ActivityItemStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  failed: 'failed',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  subtitle: string;
  valueUsd: number;
  timestamp: string;
  status: ActivityItemStatus;
  icon: string;
  walletName?: string;
}

export type ListTransactionsParams = {
/**
 * @nullable
 */
walletId?: number | null;
/**
 * @nullable
 */
type?: ListTransactionsType;
/**
 * @nullable
 */
limit?: number | null;
};

export type ListTransactionsType = typeof ListTransactionsType[keyof typeof ListTransactionsType] | null;


export const ListTransactionsType = {
  send: 'send',
  receive: 'receive',
  swap: 'swap',
  contract: 'contract',
} as const;

export type GetPortfolioHistoryParams = {
period?: GetPortfolioHistoryPeriod;
};

export type GetPortfolioHistoryPeriod = typeof GetPortfolioHistoryPeriod[keyof typeof GetPortfolioHistoryPeriod];


export const GetPortfolioHistoryPeriod = {
  '1d': '1d',
  '7d': '7d',
  '30d': '30d',
  '90d': '90d',
  '1y': '1y',
} as const;

export type GetRecentActivityParams = {
/**
 * @nullable
 */
limit?: number | null;
};