# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NP Wallet - Web3 Wallet API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemStatus } from './activityItemStatus';
import type { ActivityItemType } from './activityItemType';

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  subtitle: string;
  valueUsd: number;
  timestamp: string;
  status: ActivityItemStatus;
  icon: string;
  walletName?: string;
}