# lib/api-zod/src/generated/types/network.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NP Wallet - Web3 Wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface Network {
  id: number;
  name: string;
  symbol: string;
  chainId: number;
  rpcUrl: string;
  explorerUrl: string;
  logoUrl: string;
  isTestnet: boolean;
  color?: string;
}