# lib/api-zod/src/generated/types/sendInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NP Wallet - Web3 Wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface SendInput {
  toAddress: string;
  amount: number;
  tokenSymbol: string;
  /** @nullable */
  note?: string | null;
}