# lib/api-zod/src/generated/types/token.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NP Wallet - Web3 Wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface Token {
  id: number;
  symbol: string;
  name: string;
  priceUsd: number;
  change24hPercent: number;
  logoUrl: string;
  marketCapUsd: number;
  volume24hUsd: number;
  isPopular?: boolean;
}