# lib/api-zod/src/generated/types/walletInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * NP Wallet - Web3 Wallet API
 * OpenAPI spec version: 0.1.0
 */

export interface WalletInput {
  name: string;
  networkId: number;
  color?: string;
  /** @nullable */
  importedAddress?: string | null;
}